+++ index.js
@@ -1,3 +1,3 @@
 module.exports = function dataNavigator() {
-    console.log('data navigator has no code yet!');
+    // console.log('data navigator has no code yet!');
 };
