+++ package.json
@@ -1,14 +1,14 @@
 {
     "name": "data-navigator",
     "author": "Frank Elavsky",
-    "version": "0.0.1",
+    "version": "0.1.0",
     "keywords": [
         "visualization",
         "accessibility",
         "touch",
         "keyboard"
     ],
-    "description": "Data-navigator is a JavaScript library that allows for serial navigation of data structures using keyboard and touch inputs.",
+    "description": "Data-navigator is a JavaScript library that allows for serial navigation of data structures using a variety of input modalities and assistive technologies.",
     "scripts": {
         "build": "webpack",
         "server": "python -m http.server",
