+++ src/utilities.js
@@ -1,0 +1,9 @@
+export const describeNode = (d, descriptionOptions) => {
+    const keys = Object.keys(d);
+    let description = '';
+    keys.forEach(key => {
+        description += `${descriptionOptions.omitKeyNames ? '' : key + ': '}${d[key]}. `;
+    });
+    description += descriptionOptions.semanticLabel || 'Data point.';
+    return description;
+};
