+++ vega-lite.html
@@ -26,7 +26,7 @@
     {"@context":"https://schema.org","@type":"WebSite","description":"A tool for auditing the accessibility of data experiences.","headline":"Workbook","name":"The Chartability Workbook","url":"https://chartability.github.io/POUR-CAF/"}</script>
     <link rel="preconnect" href="https://fonts.gstatic.com">
     <link href="https://fonts.googleapis.com/css2?family=Noto+Sans:wght@400;700&amp;display=swap" rel="stylesheet"> -->
-        <link href="./src/style.css" rel="stylesheet" />
+        <link href="./examples/style.css" rel="stylesheet" />
         <style>
             /* .dn-node {
                 background: #000;
