+++ webpack.config.js
@@ -2,9 +2,13 @@ const path = require('path');
 
 module.exports = {
   entry: {
-    bundle: './src/app.js', 
-    'v-bundle': './src/vega-lite-app.js',
-    'static-bundle': './src/static-app.js'
+    'index': './src/data-navigator.js',
+    'input': './src/input.js',
+    'rendering': './src/rendering.js',
+    'structure': './src/structure.js',
+    'utilities': './src/utilities.js',
+    'v-bundle': './examples/vega-lite-app.js',
+    'static-bundle': './examples/static-app.js'
   },
   output: {
     filename: '[name].js',
